+++ Maple2.File.Parser/Xml/Table/Fish.cs
@@ -15,7 +15,7 @@ public partial class Fish : IFeatureLocale {
     [XmlAttribute] public bool fishingBook;
     [XmlAttribute] public string habitat = string.Empty;
     [XmlAttribute] public int fishMastery;
-    [XmlAttribute] public int rank;
+    [XmlAttribute] public short rank;
     [XmlAttribute] public string smallSize = string.Empty;
     [XmlAttribute] public string bigSize = string.Empty;
     [XmlAttribute] public bool ignoreSpotMastery;
